requirements.txt
# Core dependency - provides Falcon-512 post-quantum signatures
liboqs-python>=0.14.1